+++ mirrord/agent/src/main.rs
@@ -120,7 +120,6 @@ impl State {
                 // If we are in an ephemeral container, we use pid 1.
                 (true, Some(container_handle), pid)
             }
-            // if not, we use the pid of the target container or fallback to self
             cli::Mode::Targetless | cli::Mode::BlackboxTest => (false, None, "self".to_string()),
         };