+++ mirrord/config/src/lib.rs
@@ -21,6 +21,7 @@ use config::{ConfigContext, ConfigError, MirrordConfig};
 use mirrord_analytics::CollectAnalytics;
 use mirrord_config_derive::MirrordConfig;
 use schemars::JsonSchema;
+use target::Target;
 use tera::Tera;
 use tracing::warn;
 
@@ -143,7 +144,10 @@ const PAUSE_WITHOUT_STEAL_WARNING: &str =
 ///         "ignore_localhost": false,
 ///         "unix_streams": "bear.+"
 ///       },
-///       "dns": false
+///       "dns": false,
+///       "copy_target": {
+///         "scale_down": false
+///       }
 ///     },
 ///   },
 ///   "operator": true,
@@ -418,6 +422,45 @@ impl LayerConfig {
                 ))?
             }
         }
+
+        if self.feature.copy_target.enabled {
+            if !self.operator {
+                return Err(ConfigError::Conflict(
+                    "The copy target feature requires a mirrord operator, \
+                   please either disable this option or use the operator."
+                        .into(),
+                ));
+            }
+
+            if self.target.path.is_none() {
+                return Err(ConfigError::Conflict(
+                    "The copy target feature is not compatible with a targetless agent, \
+                    please either disable this option or specify a target."
+                        .into(),
+                ));
+            }
+
+            if !self.feature.network.incoming.is_steal() {
+                context.add_warning(
+                    "Using copy target feature without steal mode \
+                    may result in unreturned responses in cluster \
+                    because the underlying app instance is not copied \
+                    and therefore not running in the copied pod"
+                        .into(),
+                );
+            }
+
+            if self.feature.copy_target.scale_down
+                && !matches!(self.target.path, Some(Target::Deployment(..)))
+            {
+                return Err(ConfigError::Conflict(
+                    "The scale down feature is compatible only with deployment targets, \
+                    please either disable this option or specify a deployment target."
+                        .into(),
+                ));
+            }
+        }
+
         Ok(())
     }
 }
@@ -694,6 +737,7 @@ mod tests {
                         ..Default::default()
                     })),
                 })),
+                copy_target: None,
             }),
             connect_tcp: None,
             operator: None,