+++ mirrord/operator/src/crd.rs
@@ -88,6 +88,7 @@ pub struct MirrordOperatorSpec {
     pub features: Option<Vec<OperatorFeatures>>,
     pub license: LicenseInfoOwned,
     pub protocol_version: Option<String>,
+    pub copy_target_enabled: Option<bool>,
 }
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize, JsonSchema)]
@@ -118,7 +119,28 @@ pub struct LicenseInfoOwned {
     pub fingerprint: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize, JsonSchema)]
+#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq, Serialize, JsonSchema)]
 pub enum OperatorFeatures {
     ProxyApi,
+}
+
+/// This [`Resource`](kube::Resource) represents a copy pod created from an existing [`Target`]
+/// (operator's copy pod feature).
+#[derive(CustomResource, Clone, Debug, Deserialize, Serialize, JsonSchema)]
+#[kube(
+    group = "operator.metalbear.co",
+    version = "v1",
+    kind = "CopyTarget",
+    root = "CopyTargetCrd",
+    namespaced
+)]
+pub struct CopyTargetSpec {
+    /// Original target. Only [`Target::Pod`] and [`Target::Deployment`] are accepted.
+    pub target: Target,
+    /// How long should the operator keep this pod alive after its creation.
+    /// The pod is deleted when this timout has expired and there are no connected clients.
+    pub idle_ttl: Option<u32>,
+    /// Should the operator scale down target deployment to 0 while this pod is alive.
+    /// Ignored if [`Target`] is not [`Target::Deployment`].
+    pub scale_down: bool,
 }